requirements.txt
googletrans==4.0.0-rc1
text_to_sign
scikit-learn==1.0.2
kera
moviepy